Update scraper.gs
added function that removes duplicate rows. The spreadsheet will update if there's a change in the update clock in the source page. Nevertheless, sometimes a specific record is not updated, which leads to duplication once the data is parsed and copied. This only affects a couple of sheets: Regiao & Mapas. This functions solves that.